requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.0
django-extra-fields==3.0.2
django-heroku==0.3.1
django-jalali==6.0.0
django-jalali-date==1.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
idna==3.3
jdatetime==4.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
suds==1.0.0
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
